requirements.txt
# ── Core ─────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ── Database ─────────────────────────────────
pymongo>=4.6.0

# ── AI / LLM ────────────────────────────────
google-genai>=1.0.0

# ── Speech-to-Text ──────────────────────────
groq>=0.4.0

# ── Text-to-Speech ──────────────────────────
elevenlabs>=1.0.0
gTTS>=2.5.0

# ── WebSocket ────────────────────────────────
websockets>=12.0

